comment and do some cleaning
