+++ 1-exchanges/orderbook/src/lib.rs
@@ -2,7 +2,9 @@
 use scrypto::prelude::*;
 
 mod dex;
+mod order;
 
+//Define the Open Order book smart contrat entries.
 blueprint! {
 
     pub struct Market {
@@ -16,7 +18,15 @@ blueprint! {
     }
 
     impl Market {
-        // create a market with specified quote and base token as initiale quantity in the vault.
+        /// create a market with specified quote and base token as initial quantity in the vault.
+        /// A name identify the market.
+        /// Market owner call this method to create a new market.
+        /// Return the created market component address and market admin access badge.
+        ///
+        /// Market define 2 level of access:
+        ///  * admin: badge return by this function that can withdraw fee gain during asset transfert
+        ///  * trader or user badge needed to push order to the book
+        ///
         pub fn instantiate_market(
             quote_token: Address,
             base_token: Address,
@@ -53,6 +63,9 @@ blueprint! {
             )
         }
 
+        /// To trade, user must create a openorders badge to get access to the bid and ask method.
+        /// create_openorders create a data structure store in the market to store transferred asset during order matching.
+        /// Return the badge needed to push order and withdraw asset transferred by order execution.
         pub fn create_openorders(&mut self) -> Bucket {
             let badge = self.orders_badge_minter.authorize(|auth| {
                 self.orders_badge_def.mint_non_fungible(
@@ -64,7 +77,7 @@ blueprint! {
                 )
             });
 
-            let orders = dex::UserOrders::new(self.dex.params.quote_token.clone(), self.dex.params.base_token.clone());
+            let orders = order::UserOrders::new(self.dex.params.quote_token.clone(), self.dex.params.base_token.clone());
             self.dex.user_orders.insert(
                 badge.get_non_fungible_keys().get(0).unwrap().clone(),
                 orders,
@@ -74,8 +87,19 @@ blueprint! {
         }
 
         //use u8 for order type beacause I didn't find an example with an Option as Tx parameter.
+        /// Push a bid order to the order book. Bid order are a max limit price, a amount of base to buy and a quote bucket containing all the quote
+        /// needed to by the base at the limit price.
+        /// Order type define how the order will be matched.
+        ///  * limit: 0 when the order is push the maximum amount of quote is transferred depending on the available opposite order in the book.
+        ///    If some amount can't be matched, the remaining is added to the order book.
+        ///  * Immediate or Cancel: 1 same as limit but if a remaining amount can't be matched, it's cancelled.
+        ///  * Post: 2 the order is not matched an immediately added to the order book. It can be useful to decrease the fee (see Fee management).
+        ///
+        /// Return the order id address use to cancel it and the rest of the quote not used.
+        /// Provided quote is locked while the order is still pending and it's free when the order is cancelled or transferred when it's matched.
+        ///
         #[auth(orders_badge_def)]
-        pub fn buy_order(&mut self, price: Decimal, amount: Decimal, ordre_type: u8, quote: Bucket) -> (Bucket, Bucket) { //, auth: BucketRef
+        pub fn bid_order(&mut self, price: Decimal, amount: Decimal, ordre_type: u8, quote: Bucket) -> (Bucket, Bucket) { //, auth: BucketRef
             info!("buy_order");
             let owner_keys = auth.get_non_fungible_keys();
             let data: BadgeData = auth
@@ -91,8 +115,19 @@ blueprint! {
             )
         }
 
+        /// Push a ask order to the order book. Ask order are a min limit price, a amount of base to sell and a base bucket containing
+        /// all the base to sell.
+        /// Order type define how the order will be matched.
+        ///  * limit: 0 when the order is push the maximum amount of quote is transferred depending on the available opposite order in the book.
+        ///    If some amount can't be matched, the remaining is added to the order book.
+        ///  * Immediate or Cancel: 1 same as limit but if a remaining amount can't be matched, it's cancelled.
+        ///  * Post: 2 the order is not matched an immediately added to the order book. It can be useful to decrease the fee (see Fee management).
+        ///
+        /// Return the order id address use to cancel it and the rest of the quote not used.
+        /// Provided base is locked while the order is still pending and it's free when the order is cancelled or transferred when it's matched.
+        ///
         #[auth(orders_badge_def)]
-        pub fn sell_order(&mut self, price: Decimal, amount: Decimal, ordre_type: u8, base: Bucket) -> (Bucket, Bucket) { //, auth: BucketRef
+        pub fn ask_order(&mut self, price: Decimal, amount: Decimal, ordre_type: u8, base: Bucket) -> (Bucket, Bucket) {
             info!("sell order");
             let owner_keys = auth.get_non_fungible_keys();
             let data: BadgeData = auth
@@ -108,6 +143,9 @@ blueprint! {
             )
         }
 
+        ///
+        /// Withdraw all the asset (quote, base) store in the provided badge openorders and return it.
+        /// Locked quote or base can't be withdrawn until associated order are pending in the order book.
         #[auth(orders_badge_def)]
         pub fn withdraw(&mut self) -> (Bucket, Bucket) {
             info!("withdraw order");
@@ -123,11 +161,12 @@ blueprint! {
             .ok_or_else(|| panic!("Badge provided not declared call create_openorders to get one"))
             .unwrap();
 
-            info!("withdraw user_orders.quote_vault:{} user_orders.base_vault:{}", user_orders.quote_vault.amount(), user_orders.base_vault.amount());
-
             (user_orders.quote_vault.take_all(), user_orders.base_vault.take_all())
         }
 
+        /// Cancel the order with the specified order_id_address. The order must owned bu the auth user badge provided.
+        /// Order locked asset are transferred to the user openordersquote and base vault.
+        ///
         #[auth(orders_badge_def)]
         pub fn cancel_order(&mut self, order_id_badge: Bucket)  -> Bucket {
             info!("cancel order");
